- 实现excel文档下载

+++ src/main/java/com/es/daily_report/controllers/ReportController.java
@@ -1,13 +1,21 @@
 package com.es.daily_report.controllers;
 
 import com.alibaba.excel.EasyExcel;
+import com.alibaba.excel.ExcelWriter;
+import com.alibaba.excel.write.merge.LoopMergeStrategy;
+import com.alibaba.excel.write.metadata.WriteSheet;
+import com.alibaba.excel.write.metadata.WriteTable;
 import com.auth0.jwt.interfaces.Claim;
+import com.es.daily_report.dao.ProductDao;
+import com.es.daily_report.dao.ProjectDao;
 import com.es.daily_report.dao.ReportDao;
 import com.es.daily_report.dao.TaskDao;
 
 import com.es.daily_report.dto.ReportQuery;
 
 import com.es.daily_report.dto.UserInfoDTO;
+import com.es.daily_report.entities.Product;
+import com.es.daily_report.entities.Project;
 import com.es.daily_report.entities.Report;
 import com.es.daily_report.entities.Task;
 
@@ -17,10 +25,14 @@ import com.es.daily_report.mapstruct.TaskVoMapper;
 import com.es.daily_report.services.WebService;
 import com.es.daily_report.shiro.JwtUtil;
 import com.es.daily_report.utils.Result;
+import com.es.daily_report.vo.ExcelVO;
 import com.es.daily_report.vo.ReportVO;
 import com.es.daily_report.vo.TaskVO;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shiro.authz.annotation.Logical;
 import org.apache.shiro.authz.annotation.RequiresRoles;
@@ -35,9 +47,7 @@ import java.io.IOException;
 import java.net.URLEncoder;
 import java.text.DateFormat;
 import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -53,6 +63,12 @@ public class ReportController {
 
     @Autowired
     private TaskVoMapper taskVoMapper;
+
+    @Autowired
+    private ProductDao productDao;
+
+    @Autowired
+    private ProjectDao projectDao;
 
     private ObjectMapper objectMapper;
 
@@ -107,6 +123,10 @@ public class ReportController {
         if (username == null) {
             return Result.failure(ErrorType.TOKEN_INVALID);
         }
+        String department = departmentFromToken(token);
+        if (department == null) {
+            return Result.failure(ErrorType.TOKEN_INVALID);
+        }
         Report report = reportDao.queryOnDay(account, reportVO.getOnDay());
         if (report != null) { // remove origin report, override it
             taskDao.removeTasksOfReport(report.getId());
@@ -115,6 +135,7 @@ public class ReportController {
         report = Report.builder()
                 .workCode(account)
                 .authorName(username)
+                .department(department)
                 .onDay(reportVO.getOnDay())
                 .status(reportVO.getStatus())
                 .committed(new Date())
@@ -151,12 +172,10 @@ public class ReportController {
         }).collect(Collectors.toList());
     }
 
-    @GetMapping
-    public Result<?> query(@RequestParam("type") String type,
-                           @RequestParam("condition") String content,
-                           @RequestParam("from") String from,
-                           @RequestParam("to") String to) {
-
+    List<Report> reportsByCondition(String type,
+                                    String content,
+                                    String from,
+                                    String to) {
         List<String> staffNumbers = new ArrayList<>();
 
         switch (type) {
@@ -181,8 +200,11 @@ public class ReportController {
                     staffNumbers.add(staff.getWorkcode());
                 }
                 break;
+            case "4":
+                List<String> projectMembers = projectDao.queryMemberNumber(content);
+                staffNumbers.addAll(projectMembers);
         }
-        List<Report> reports = new ArrayList<Report>();
+        List<Report> reports = new ArrayList<>();
         try {
             final Date fromDate = DateFormat.getDateInstance().parse(from);
             final Date toDate = DateFormat.getDateInstance().parse(to);
@@ -197,72 +219,80 @@ public class ReportController {
         } catch (ParseException e) {
             e.printStackTrace();
         }
+        return reports;
+    }
 
+    @GetMapping
+    public Result<?> query(@RequestParam("type") String type,
+                           @RequestParam("condition") String content,
+                           @RequestParam("from") String from,
+                           @RequestParam("to") String to) {
+        List<Report> reports = reportsByCondition(type, content, from, to);
         return Result.success(reportsWithTasks(reports));
     }
 
     //TODO: 组织任务和日志
     @GetMapping("/download")
+    @ApiOperation("根据条件下载符合要求的员工日志表格")
+    @ApiImplicitParams({
+        @ApiImplicitParam(name="type",value="查询类型0: 用户编号 1: 部门编号 2: 公司编号 3: 职位编号 4: 项目编号",dataType="string", paramType = "query"),
+        @ApiImplicitParam(name="condition",value="查询条件",dataType="string", paramType = "query"),
+            @ApiImplicitParam(name="from",value="起始日期：\"yyyy-MM-dd hh:mm:ss\"",dataType="string", paramType = "query"),
+            @ApiImplicitParam(name="to",value="截止日期：\"yyyy-MM-dd hh:mm:ss\"",dataType="string", paramType = "query")
+    })
     public void download(@RequestParam("type") String type,
                          @RequestParam("condition") String content,
                          @RequestParam("from") String from,
                          @RequestParam("to") String to,
                          HttpServletResponse response) throws IOException {
-        List<String> staffNumbers = new ArrayList<>();
-        String sheetName = "文鼎创";
-        switch (type) {
-            case "0":
-                UserInfoDTO targetStaff = webService.getUserInfoByWorkCode(content);
-                staffNumbers.add(targetStaff.getWorkcode());
-                sheetName = targetStaff.getLastname();
-                break;
-            case "1":
-                UserInfoDTO[] staffsInDepartment = webService.getUserInfoByDepartmentId(content);
-                for (UserInfoDTO staff : staffsInDepartment) {
-                    staffNumbers.add(staff.getWorkcode());
-                }
-                if (staffsInDepartment.length > 0) {
-                    sheetName = staffsInDepartment[0].getDepartmentname();
-                }
-                break;
-            case "2":
-                UserInfoDTO[] staffsInCompany = webService.getUserInfoByCompany(content);
-                for (UserInfoDTO staff : staffsInCompany) {
-                    staffNumbers.add(staff.getWorkcode());
-                }
-                sheetName = "文鼎创";
-                break;
-            case "3":
-                UserInfoDTO[] staffInJobTitle = webService.getUserInfoByJobTitleId(content);
-                for (UserInfoDTO staff : staffInJobTitle) {
-                    staffNumbers.add(staff.getWorkcode());
-                }
-                if (staffInJobTitle.length > 0) {
-                    sheetName = staffInJobTitle[0].getJobtitle();
-                }
-                break;
-        }
-        List<Report> reports = new ArrayList<Report>();
-        try {
-            final Date fromDate = DateFormat.getDateInstance().parse(from);
-            final Date toDate = DateFormat.getDateInstance().parse(to);
-            staffNumbers.forEach(workCode -> {
-                ReportQuery reportQuery = ReportQuery.builder()
-                        .staffNo(workCode)
-                        .start(fromDate)
-                        .end(toDate)
-                        .build();
-                reports.addAll(reportDao.listByRange(reportQuery));
-            });
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
         // 这里注意 有同学反应使用swagger 会导致各种问题，请直接用浏览器或者用postman
         response.setContentType("application/vnd.ms-excel");
         response.setCharacterEncoding("utf-8");
-        // 这里URLEncoder.encode可以防止中文乱码 当然和easyexcel没有关系
-        String fileName = URLEncoder.encode("工作日志", "UTF-8");
+        String fileName = URLEncoder.encode(content, "UTF-8");
         response.setHeader("Content-disposition", "attachment;filename=" + fileName + ".xlsx");
-        EasyExcel.write(response.getOutputStream(), Report.class).sheet(sheetName).doWrite(reportsWithTasks(reports));
+
+        List<Report> reports = reportsByCondition(type, content, from, to);
+
+        Map<String, List<ExcelVO>> excel= new HashMap<>();
+        reports.forEach(report -> {
+            List<Task> tasks = taskDao.queryByReport(report.getId());
+            tasks.forEach(task -> {
+                if (type.equals("4") && !task.getProjectId().equals(content)) {
+                    return;
+                }
+                Project project = projectDao.queryByNumber(task.getProjectId());
+                Product product = productDao.queryByNumber(task.getProductId());
+                ExcelVO excelVO = ExcelVO.builder()
+                        .staffName(report.getAuthorName())
+                        .department(report.getDepartment())
+                        .reportDate(DateFormat.getDateInstance().format(report.getOnDay()))
+                        .commitDate(DateFormat.getDateInstance().format(report.getCommitted()))
+                        .staffNo(report.getWorkCode())
+                        .taskName(task.getName())
+                        .taskCost(task.getCost().toString())
+                        .taskDetail(task.getDetails())
+                        .productNo(task.getProductId())
+                        .projectNo(task.getProjectId())
+                        .productName(product.getName())
+                        .projectName(project.getName())
+                        .productLine(product.getInLine())
+                        .build();
+                List<ExcelVO> sheet = excel.get(report.getWorkCode() + "-" + report.getAuthorName());
+                if (sheet == null) {
+                    sheet = new ArrayList<>();
+                    sheet.add(excelVO);
+                    excel.put(report.getWorkCode() + "-" + report.getAuthorName(), sheet);
+                } else {
+                    sheet.add(excelVO);
+                }
+            });
+        });
+        ExcelWriter excelWriter = EasyExcel.write(response.getOutputStream()).build();
+        excel.forEach((key, value) -> {
+                WriteSheet writeSheet = EasyExcel.writerSheet(key).build();
+                WriteTable writeTable = EasyExcel.writerTable().head(ExcelVO.class).needHead(true).build();
+                excelWriter.write(value, writeSheet, writeTable);
+        });
+        excelWriter.finish();
     }
 }
